Refactory/v7

+++ webapi/Controllers/Account/Edit/2FaController.cs
@@ -2,8 +2,8 @@
 using Microsoft.AspNetCore.Mvc;
 using webapi.DTO;
 using webapi.Exceptions;
+using webapi.Interfaces;
 using webapi.Interfaces.Services;
-using webapi.Interfaces.SQL;
 using webapi.Localization;
 using webapi.Models;
 using webapi.Services;
@@ -18,9 +18,8 @@ namespace webapi.Controllers.Account.Edit
     {
         private const string CODE = "2FaCode";
 
-        private readonly IUpdate<UserModel> _updateUser;
-        private readonly IRead<UserModel> _readUser;
-        private readonly ICreate<NotificationModel> _createNotification;
+        private readonly IRepository<UserModel> _userRepository;
+        private readonly IRepository<NotificationModel> _notificationRepository;
         private readonly IPasswordManager _passwordManager;
         private readonly IUserInfo _userInfo;
         private readonly ITokenService _tokenService;
@@ -30,9 +29,8 @@ namespace webapi.Controllers.Account.Edit
         private readonly IValidation _validation;
 
         public _2FaController(
-            IUpdate<UserModel> updateUser,
-            IRead<UserModel> readUser,
-            ICreate<NotificationModel> createNotification,
+            IRepository<UserModel> userRepository,
+            IRepository<NotificationModel> notificationRepository,
             IPasswordManager passwordManager,
             IUserInfo userInfo,
             ITokenService tokenService,
@@ -41,9 +39,8 @@ namespace webapi.Controllers.Account.Edit
             IEmailSender emailSender,
             IValidation validation)
         {
-            _updateUser = updateUser;
-            _readUser = readUser;
-            _createNotification = createNotification;
+            _userRepository = userRepository;
+            _notificationRepository = notificationRepository;
             _passwordManager = passwordManager;
             _userInfo = userInfo;
             _tokenService = tokenService;
@@ -56,34 +53,31 @@ namespace webapi.Controllers.Account.Edit
         [HttpPost("start")]
         public async Task<IActionResult> SendVerificationCode([FromQuery] string password)
         {
-            try
-            {
-                var user = await _readUser.ReadById(_userInfo.UserId, null);
-
-                bool IsCorrect = _passwordManager.CheckPassword(password, user.password);
-                if (!IsCorrect)
-                    return StatusCode(401, new { message = AccountErrorMessage.PasswordIncorrect });
-
-                int code = _generateCode.GenerateSixDigitCode();
-
-                await _emailSender.SendMessage(new EmailDto
-                {
-                    username = _userInfo.Username,
-                    email = _userInfo.Email,
-                    subject = EmailMessage.Change2FaHeader,
-                    message = EmailMessage.Change2FaBody + code
-                });
-
-                HttpContext.Session.SetString(CODE, code.ToString());
-
-                return StatusCode(200);
-            }
-            catch (UserException)
+            var user = await _userRepository.GetById(_userInfo.UserId);
+            if (user is null)
             {
                 _tokenService.DeleteTokens();
                 _logger.LogWarning("Tokens was deleted");
                 return StatusCode(404);
             }
+
+            bool IsCorrect = _passwordManager.CheckPassword(password, user.password);
+            if (!IsCorrect)
+                return StatusCode(401, new { message = AccountErrorMessage.PasswordIncorrect });
+
+            int code = _generateCode.GenerateSixDigitCode();
+
+            await _emailSender.SendMessage(new EmailDto
+            {
+                username = _userInfo.Username,
+                email = _userInfo.Email,
+                subject = EmailMessage.Change2FaHeader,
+                message = EmailMessage.Change2FaBody + code
+            });
+
+            HttpContext.Session.SetString(CODE, code.ToString());
+
+            return StatusCode(200);
         }
 
         [HttpPut("confirm/{enable}")]
@@ -99,18 +93,25 @@ namespace webapi.Controllers.Account.Edit
                 if (!code.Equals(correctCode))
                     return StatusCode(422, new { message = AccountErrorMessage.CodeIncorrect });
 
-                var user = await _readUser.ReadById(_userInfo.UserId, null);
+                var user = await _userRepository.GetById(_userInfo.UserId);
+                if (user is null)
+                {
+                    _tokenService.DeleteTokens();
+                    _logger.LogWarning("Tokens was deleted");
+                    return StatusCode(404);
+                }
+
                 if (user.is_2fa_enabled == enable)
                     return StatusCode(409);
 
                 user.is_2fa_enabled = enable;
-                await _updateUser.Update(user, null);
+                await _userRepository.Update(user);
 
                 string message = enable
                     ? "Your two-factor authentication status has been successfully updated! Your account is now even more secure. Thank you for prioritizing security with us."
                     : "Your two-factor authentication has been disabled. Please ensure that you take additional precautions to secure your account. If this change was not authorized, contact our support team immediately. Thank you for staying vigilant about your account security.";
 
-                await _createNotification.Create(new NotificationModel
+                await _notificationRepository.Add(new NotificationModel
                 {
                     message_header = $"Someone changed changed your 2FA status",
                     message = message,
@@ -119,15 +120,18 @@ namespace webapi.Controllers.Account.Edit
                     is_checked = false,
                     receiver_id = _userInfo.UserId
                 });
+
                 HttpContext.Session.SetString(Constants.CACHE_USER_DATA, true.ToString());
 
                 return StatusCode(200);
             }
-            catch (UserException)
+            catch (EntityNotCreatedException ex)
             {
-                _tokenService.DeleteTokens();
-                _logger.LogWarning("Tokens was deleted");
-                return StatusCode(404);
+                return StatusCode(500, new { message = ex.Message });
+            }
+            catch (EntityNotUpdatedException ex)
+            {
+                return StatusCode(500, new { message = ex.Message });
             }
         }
     }
